Refactored with separation of concerns and imports/exports.
