--- frontend/src/js/page-simulation.js
@@ -1,1 +1,0 @@
-import '../scss/page-simulation.scss'
