+++ frontend/src/js/page-simulation/helpers/constants.js
@@ -1,0 +1,2 @@
+export const IOF = 6.38 / 100
+export const INTEREST_RATE = 2.34 / 100
